Update LINUX.md to reflect current bug in current openssl-client / remote -ssh interaction
Currently, the suggested ssh config file configuration does not work, because it contains unescaped spaces. Unless a fix is deployed by the developers of the remote-ssh extension, hostnames with spaces make it unable to connect to the host, because an invalid character error is raised somewhere. 
See the following issue on the extension page: https://github.com/microsoft/vscode-remote-release/issues/9370 

I had the issue appear for me in the bootcamp yesterday and it took me a good 6 hours to find the fix. 

Assuming that it might take Microsoft a while to deploy a fix, I suggest changing the suggested naming for the host to not contain any spaces.